OCPBUGS-44103: disable go workspace and pin distribution/distribution/v3
The distribution/distribution/v3 library needs to be pinned to a version
before the `ocischema.DeserializedImageIndex` type as introduced. That
is because `oc` and `library-go` cannot currently handle that type and
nor do we know when it will.

Due to way go workspace works, we could not downgrade the distribution
version we used in v1 because we need a more recent version for v2
(needed by containers/image). Removing the workspace allows us to have
distinct dependency versions thus making it possible to have distinct
development cadences for v1 and v2.

To make it possible for users to still invoke --v1 or --v2 from the same
binary, we use the `//go:embed` directive to embed the v2 binary into
the v1 oc-mirror, extract and execute the binary at runtime. It incurs a
small time penalty but it has the same affect as importing the v2 module
into v1 without the downsides.